refactor(w1): extract PluginSettings to domain/settings (#99)
PluginSettings + DEFAULT_SETTINGS are domain configuration, not bridge
mechanics. Move to src/domain/settings/PluginSettings.ts and re-export
from the old IBridge location to keep callers green during migration.
