+++ src/domain/settings/PluginSettings.ts
@@ -1,0 +1,23 @@
+/**
+ * Domain-level plugin configuration. Persisted via SettingsPort and
+ * read by use cases that need to resolve vault paths or behaviour flags.
+ */
+export interface PluginSettings {
+	readonly locale: string
+	readonly specsFolder: string
+	readonly archiveFolder: string
+	readonly decisionsFolder: string
+	readonly constitutionFile: string
+	readonly gateStrictness: 'strict' | 'lenient'
+	readonly teamMode: boolean
+}
+
+export const DEFAULT_SETTINGS: PluginSettings = {
+	locale: 'en',
+	specsFolder: 'specs',
+	archiveFolder: 'archive',
+	decisionsFolder: 'decisions',
+	constitutionFile: 'CONSTITUTION.md',
+	gateStrictness: 'strict',
+	teamMode: false,
+}
